Rename pippedsdk to sdk
